requirements.txt
--extra-index-url https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple

aiodns==3.0.0; python_version >= "3.8" and python_version < "4.0"
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.8" and python_version < "4.0"
arclet-alconna==0.4.3; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2; python_version >= "3.8" and python_version < "4.0"
attrs==21.4.0; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.5.0"
brotli==1.0.9; python_version >= "3.8" and python_version < "4.0"
cchardet==2.1.7; python_version >= "3.8" and python_version < "4.0"
cffi==1.15.0; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==2.0.12; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.5.0"
colorama==0.4.4; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32" and python_full_version >= "3.5.0"
croniter==1.3.4; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.4.0"
frozenlist==1.3.0; python_version >= "3.8" and python_version < "4.0"
graia-ariadne==0.5.3.post3; python_version >= "3.8" and python_version < "4.0"
graia-broadcast==0.15.6; python_version >= "3.8" and python_version < "4.0"
graia-saya==0.0.14; python_version >= "3.8" and python_version < "4.0"
graia-scheduler==0.0.6; python_version >= "3.8" and python_version < "4.0"
graiax-silkcoder==0.1.6; python_version >= "3.6"
idna==3.3; python_version >= "3.8" and python_version < "4.0"
imageio-ffmpeg==0.4.5; python_version >= "3.6"
loguru==0.6.0; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.2; python_version >= "3.8" and python_version < "4.0"
prompt-toolkit==3.0.28; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.2"
pycares==4.1.2; python_version >= "3.8" and python_version < "4.0"
pycparser==2.21; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.4.0"
pydantic==1.9.0; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.1"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.4.0"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.4.0"
typing-extensions==4.1.1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.1"
wcwidth==0.2.5; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.2"
win32-setctime==1.1.0; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
yarl==1.7.2; python_version >= "3.8" and python_version < "4.0"
